backend/requirements.txt
fastapi
uvicorn[standard]
numpy
sqlalchemy
python-jose[cryptography]
python-multipart
websockets
pydantic
python-dotenv
